Refactored the ISR to not have floating point division. Changed how it updates the time that is sent to the device. Moved around some stuff from prj.conf and boards/*.conf.

+++ workspace/open_rowing_monitor_esp32/modules/hardware_driver/GpioTimerService/GpioTimerService.h
@@ -28,7 +28,7 @@ private:
 
     // IPC: Message Queue
     struct k_msgq impulseQueue;
-    char __aligned(8) impulseQueueBuffer[IMPULSE_QUEUE_SIZE * sizeof(double)];
+    char __aligned(8) impulseQueueBuffer[IMPULSE_QUEUE_SIZE * sizeof(uint32_t)];
 
     // THREAD DATA
     // We keep the struct here, but the STACK will be defined in the .cpp file
